refactor: LottoResultAnalyst 추상화 해제
