+++ src/main/kotlin/lotto/view/LottoController.kt
@@ -1,7 +1,6 @@
 package lotto.view
 
 import lotto.domain.analysis.LottoAnalysisRequest
-import lotto.domain.analysis.LottoResultAnalyst
 import lotto.domain.shop.LottoShop
 
 class LottoController(
